requirements.txt
pandas
selenium
webdriver-manager
beautifulsoup4
requests
pillow
boto3
